+++ app/src/main/java/com/averycorp/prismtask/data/local/database/PrismTaskDatabase.kt
@@ -14,6 +14,8 @@ import com.averycorp.prismtask.data.local.dao.HabitDao
 import com.averycorp.prismtask.data.local.dao.HabitLogDao
 import com.averycorp.prismtask.data.local.dao.HabitTemplateDao
 import com.averycorp.prismtask.data.local.dao.LeisureDao
+import com.averycorp.prismtask.data.local.dao.MedicationDao
+import com.averycorp.prismtask.data.local.dao.MedicationDoseDao
 import com.averycorp.prismtask.data.local.dao.MedicationRefillDao
 import com.averycorp.prismtask.data.local.dao.MilestoneDao
 import com.averycorp.prismtask.data.local.dao.MoodEnergyLogDao
@@ -46,6 +48,8 @@ import com.averycorp.prismtask.data.local.entity.HabitEntity
 import com.averycorp.prismtask.data.local.entity.HabitLogEntity
 import com.averycorp.prismtask.data.local.entity.HabitTemplateEntity
 import com.averycorp.prismtask.data.local.entity.LeisureLogEntity
+import com.averycorp.prismtask.data.local.entity.MedicationDoseEntity
+import com.averycorp.prismtask.data.local.entity.MedicationEntity
 import com.averycorp.prismtask.data.local.entity.MedicationRefillEntity
 import com.averycorp.prismtask.data.local.entity.MilestoneEntity
 import com.averycorp.prismtask.data.local.entity.MoodEnergyLogEntity
@@ -101,9 +105,11 @@ import com.averycorp.prismtask.data.local.entity.WeeklyReviewEntity
         TaskCompletionEntity::class,
         FocusReleaseLogEntity::class,
         DailyEssentialSlotCompletionEntity::class,
-        MilestoneEntity::class
+        MilestoneEntity::class,
+        MedicationEntity::class,
+        MedicationDoseEntity::class
     ],
-    version = 53,
+    version = 54,
     exportSchema = false
 )
 abstract class PrismTaskDatabase : RoomDatabase() {
@@ -164,4 +170,8 @@ abstract class PrismTaskDatabase : RoomDatabase() {
     abstract fun dailyEssentialSlotCompletionDao(): DailyEssentialSlotCompletionDao
 
     abstract fun milestoneDao(): MilestoneDao
+
+    abstract fun medicationDao(): MedicationDao
+
+    abstract fun medicationDoseDao(): MedicationDoseDao
 }